refactor: apis/game 폴더구조 변경
