+++ src/components/features/gamble/match-on.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from 'react';
 import PredictCard from '../home/predict-card';
-import { GameDto, GetGamesRequest } from '@/services/apis/game/dto';
-import { getGames } from '@/services/apis/game';
+import { GameDto, GetGamesRequest } from '@/services/apis/game/game.type';
+import { getGames } from '@/services/apis/game/game.api';
 import { formatFromTo } from '@/lib/utils';
 import NoGameCard from './no-game-card';
 import MatchPredictionCalendar from '@/components/common/calendar';
