+++ src/components/features/gamble/predict-result.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from 'react';
 import PredictCard from '../home/predict-card';
-import { GameDto, GetMyPredictionsRequest } from '@/services/apis/game/dto';
-import { getMyPredictions } from '@/services/apis/game';
+import { GameDto, GetMyPredictionsRequest } from '@/services/apis/game/game.type';
+import { getMyPredictions } from '@/services/apis/game/game.api';
 import { formatFromTo } from '@/lib/utils';
 import NoGameCard from './no-game-card';
 import Stat from './stat';
